benchmark array test step2
